Extract topics to separate routes file
